specialization ChainStateN bugfix ini init
